@@ -12,7 +12,7 @@ class FieldsetLayout extends Layout {
 	 * @param array $config Configuration options
 	 * @param FieldLayout[] $config['items'] Items to add
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Parent constructor
 		parent::__construct( $config );
 
@@ -23,7 +23,7 @@ class FieldsetLayout extends Layout {
 
 		// Initialization
 		$this
-			->addClasses( array( 'oo-ui-fieldsetLayout' ) )
+			->addClasses( [ 'oo-ui-fieldsetLayout' ] )
 			->prependContent( $this->icon, $this->label, $this->group );
 		if ( isset( $config['items'] ) ) {
 			$this->addItems( $config['items'] );
