@@ -11,15 +11,15 @@ class HorizontalLayout extends Layout {
 	 * @param array $config Configuration options
 	 * @param Widget[]|Layout[] $config['items'] Widgets or other layouts to add to the layout.
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Parent constructor
 		parent::__construct( $config );
 
 		// Mixins
-		$this->mixin( new GroupElement( $this, array_merge( $config, array( 'group' => $this ) ) ) );
+		$this->mixin( new GroupElement( $this, array_merge( $config, [ 'group' => $this ] ) ) );
 
 		// Initialization
-		$this->addClasses( array( 'oo-ui-horizontalLayout' ) );
+		$this->addClasses( [ 'oo-ui-horizontalLayout' ] );
 		if ( isset( $config['items'] ) ) {
 			$this->addItems( $config['items'] );
 		}
