@@ -16,7 +16,7 @@ class FlaggedElement extends ElementMixin {
 	 *
 	 * @var string[]
 	 */
-	protected $flags = array();
+	protected $flags = [];
 
 	public static $targetPropertyName = 'flagged';
 
@@ -26,7 +26,7 @@ class FlaggedElement extends ElementMixin {
 	 * @param string|string[] $config['flags'] Flags describing importance and functionality, e.g.
 	 *   'primary', 'safe', 'progressive', 'destructive' or 'constructive'
 	 */
-	public function __construct( Element $element, array $config = array() ) {
+	public function __construct( Element $element, array $config = [] ) {
 		// Parent constructor
 		$target = isset( $config['flagged'] ) ? $config['flagged'] : $element;
 		parent::__construct( $element, $target, $config );
@@ -60,7 +60,7 @@ class FlaggedElement extends ElementMixin {
 	 * @return $this
 	 */
 	public function clearFlags() {
-		$remove = array();
+		$remove = [];
 		$classPrefix = 'oo-ui-flaggedElement-';
 
 		foreach ( $this->flags as $flag ) {
@@ -68,7 +68,7 @@ class FlaggedElement extends ElementMixin {
 		}
 
 		$this->target->removeClasses( $remove );
-		$this->flags = array();
+		$this->flags = [];
 
 		return $this;
 	}
@@ -81,8 +81,8 @@ class FlaggedElement extends ElementMixin {
 	 * @return $this
 	 */
 	public function setFlags( $flags ) {
-		$add = array();
-		$remove = array();
+		$add = [];
+		$remove = [];
 		$classPrefix = 'oo-ui-flaggedElement-';
 
 		if ( is_string( $flags ) ) {
