@@ -22,14 +22,14 @@ class LabelElement extends ElementMixin {
 	 * @param array $config Configuration options
 	 * @param string|HtmlSnippet $config['label'] Label text
 	 */
-	public function __construct( Element $element, array $config = array() ) {
+	public function __construct( Element $element, array $config = [] ) {
 		// Parent constructor
 		// FIXME 'labelElement' is a very stupid way to call '$label'
 		$target = isset( $config['labelElement'] ) ? $config['labelElement'] : new Tag( 'span' );
 		parent::__construct( $element, $target, $config );
 
 		// Initialization
-		$this->target->addClasses( array( 'oo-ui-labelElement-label' ) );
+		$this->target->addClasses( [ 'oo-ui-labelElement-label' ] );
 		$this->setLabel( isset( $config['label'] ) ? $config['label'] : null );
 	}
 
@@ -54,7 +54,7 @@ class LabelElement extends ElementMixin {
 			}
 		}
 
-		$this->element->toggleClasses( array( 'oo-ui-labelElement' ), !!$this->label );
+		$this->element->toggleClasses( [ 'oo-ui-labelElement' ], !!$this->label );
 
 		return $this;
 	}
