@@ -36,7 +36,7 @@ class ButtonWidget extends Widget {
 	 * @param string $config['target'] Target to open hyperlink in
 	 * @param boolean $config['noFollow'] Search engine traversal hint (default: true)
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Parent constructor
 		parent::__construct( $config );
 
@@ -46,17 +46,17 @@ class ButtonWidget extends Widget {
 		$this->mixin( new IndicatorElement( $this, $config ) );
 		$this->mixin( new LabelElement( $this, $config ) );
 		$this->mixin( new TitledElement( $this,
-			array_merge( $config, array( 'titled' => $this->button ) ) ) );
+			array_merge( $config, [ 'titled' => $this->button ] ) ) );
 		$this->mixin( new FlaggedElement( $this, $config ) );
 		$this->mixin( new TabIndexedElement( $this,
-			array_merge( $config, array( 'tabIndexed' => $this->button ) ) ) );
+			array_merge( $config, [ 'tabIndexed' => $this->button ] ) ) );
 		$this->mixin( new AccessKeyedElement( $this,
-			array_merge( $config, array( 'accessKeyed' => $this->button ) ) ) );
+			array_merge( $config, [ 'accessKeyed' => $this->button ] ) ) );
 
 		// Initialization
 		$this->button->appendContent( $this->icon, $this->label, $this->indicator );
 		$this
-			->addClasses( array( 'oo-ui-buttonWidget' ) )
+			->addClasses( [ 'oo-ui-buttonWidget' ] )
 			->appendContent( $this->button );
 
 		$this->setHref( isset( $config['href'] ) ? $config['href'] : null );
@@ -113,9 +113,9 @@ class ButtonWidget extends Widget {
 	 */
 	public function updateHref() {
 		if ( $this->href !== null && !$this->isDisabled() ) {
-			$this->button->setAttributes( array( 'href' => $this->href ) );
+			$this->button->setAttributes( [ 'href' => $this->href ] );
 		} else {
-			$this->button->removeAttributes( array( 'href' ) );
+			$this->button->removeAttributes( [ 'href' ] );
 		}
 		return $this;
 	}
@@ -130,9 +130,9 @@ class ButtonWidget extends Widget {
 		$this->target = is_string( $target ) ? $target : null;
 
 		if ( $this->target !== null ) {
-			$this->button->setAttributes( array( 'target' => $target ) );
+			$this->button->setAttributes( [ 'target' => $target ] );
 		} else {
-			$this->button->removeAttributes( array( 'target' ) );
+			$this->button->removeAttributes( [ 'target' ] );
 		}
 
 		return $this;
@@ -148,9 +148,9 @@ class ButtonWidget extends Widget {
 		$this->noFollow = is_bool( $noFollow ) ? $noFollow : true;
 
 		if ( $this->noFollow ) {
-			$this->button->setAttributes( array( 'rel' => 'nofollow' ) );
+			$this->button->setAttributes( [ 'rel' => 'nofollow' ] );
 		} else {
-			$this->button->removeAttributes( array( 'rel' ) );
+			$this->button->removeAttributes( [ 'rel' ] );
 		}
 
 		return $this;
