@@ -21,12 +21,12 @@ class CheckboxInputWidget extends InputWidget {
 	 * @param boolean $config['selected'] Whether the checkbox is initially selected
 	 *   (default: false)
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Parent constructor
 		parent::__construct( $config );
 
 		// Initialization
-		$this->addClasses( array( 'oo-ui-checkboxInputWidget' ) );
+		$this->addClasses( [ 'oo-ui-checkboxInputWidget' ] );
 		// Required for pretty styling in MediaWiki theme
 		$this->appendContent( new Tag( 'span' ) );
 		$this->setSelected( isset( $config['selected'] ) ? $config['selected'] : false );
@@ -34,7 +34,7 @@ class CheckboxInputWidget extends InputWidget {
 
 	protected function getInputElement( $config ) {
 		$input = new Tag( 'input' );
-		$input->setAttributes( array( 'type' => 'checkbox' ) );
+		$input->setAttributes( [ 'type' => 'checkbox' ] );
 		return $input;
 	}
 
@@ -47,9 +47,9 @@ class CheckboxInputWidget extends InputWidget {
 	public function setSelected( $state ) {
 		$this->selected = (bool)$state;
 		if ( $this->selected ) {
-			$this->input->setAttributes( array( 'checked' => 'checked' ) );
+			$this->input->setAttributes( [ 'checked' => 'checked' ] );
 		} else {
-			$this->input->removeAttributes( array( 'checked' ) );
+			$this->input->removeAttributes( [ 'checked' ] );
 		}
 		return $this;
 	}
