@@ -16,19 +16,19 @@ class IconWidget extends Widget {
 	/**
 	 * @param array $config Configuration options
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Parent constructor
 		parent::__construct( $config );
 
 		// Mixins
 		$this->mixin( new IconElement( $this,
-			array_merge( $config, array( 'iconElement' => $this ) ) ) );
+			array_merge( $config, [ 'iconElement' => $this ] ) ) );
 		$this->mixin( new TitledElement( $this,
-			array_merge( $config, array( 'titled' => $this ) ) ) );
+			array_merge( $config, [ 'titled' => $this ] ) ) );
 		$this->mixin( new FlaggedElement( $this,
-			array_merge( $config, array( 'flagged' => $this ) ) ) );
+			array_merge( $config, [ 'flagged' => $this ] ) ) );
 
 		// Initialization
-		$this->addClasses( array( 'oo-ui-iconWidget' ) );
+		$this->addClasses( [ 'oo-ui-iconWidget' ] );
 	}
 }
