@@ -50,15 +50,15 @@ class TextInputWidget extends InputWidget {
 	 * @param boolean $config['autocomplete'] If the field should support autocomplete
 	 *   or not (default: true)
 	 */
-	public function __construct( array $config = array() ) {
+	public function __construct( array $config = [] ) {
 		// Config initialization
-		$config = array_merge( array(
+		$config = array_merge( [
 			'type' => 'text',
 			'readOnly' => false,
 			'autofocus' => false,
 			'required' => false,
 			'autocomplete' => true,
-		), $config );
+		], $config );
 		if ( $config['type'] === 'search' ) {
 			if ( !array_key_exists( 'icon', $config ) ) {
 				$config['icon'] = 'search';
@@ -83,26 +83,26 @@ class TextInputWidget extends InputWidget {
 
 		// Initialization
 		$this
-			->addClasses( array( 'oo-ui-textInputWidget', 'oo-ui-textInputWidget-type-' . $this->type ) )
+			->addClasses( [ 'oo-ui-textInputWidget', 'oo-ui-textInputWidget-type-' . $this->type ] )
 			->appendContent( $this->icon, $this->indicator );
 		$this->setReadOnly( $config['readOnly'] );
 		if ( isset( $config['placeholder'] ) ) {
-			$this->input->setAttributes( array( 'placeholder' => $config['placeholder'] ) );
+			$this->input->setAttributes( [ 'placeholder' => $config['placeholder'] ] );
 		}
 		if ( isset( $config['maxLength'] ) ) {
-			$this->input->setAttributes( array( 'maxlength' => $config['maxLength'] ) );
+			$this->input->setAttributes( [ 'maxlength' => $config['maxLength'] ] );
 		}
 		if ( $config['autofocus'] ) {
-			$this->input->setAttributes( array( 'autofocus' => 'autofocus' ) );
+			$this->input->setAttributes( [ 'autofocus' => 'autofocus' ] );
 		}
 		if ( $config['required'] ) {
-			$this->input->setAttributes( array( 'required' => 'required', 'aria-required' => 'true' ) );
+			$this->input->setAttributes( [ 'required' => 'required', 'aria-required' => 'true' ] );
 		}
 		if ( !$config['autocomplete'] ) {
-			$this->input->setAttributes( array( 'autocomplete' => 'off' ) );
+			$this->input->setAttributes( [ 'autocomplete' => 'off' ] );
 		}
 		if ( $this->multiline && isset( $config['rows'] ) ) {
-			$this->input->setAttributes( array( 'rows' => $config['rows'] ) );
+			$this->input->setAttributes( [ 'rows' => $config['rows'] ] );
 		}
 	}
 
@@ -125,9 +125,9 @@ class TextInputWidget extends InputWidget {
 	public function setReadOnly( $state ) {
 		$this->readOnly = (bool)$state;
 		if ( $this->readOnly ) {
-			$this->input->setAttributes( array( 'readonly' => 'readonly' ) );
+			$this->input->setAttributes( [ 'readonly' => 'readonly' ] );
 		} else {
-			$this->input->removeAttributes( array( 'readonly' ) );
+			$this->input->removeAttributes( [ 'readonly' ] );
 		}
 		return $this;
 	}
@@ -137,7 +137,7 @@ class TextInputWidget extends InputWidget {
 			return new Tag( 'textarea' );
 		} else {
 			$input = new Tag( 'input' );
-			$input->setAttributes( array( 'type' => $this->getSaneType( $config ) ) );
+			$input->setAttributes( [ 'type' => $this->getSaneType( $config ) ] );
 			return $input;
 		}
 	}
@@ -146,7 +146,7 @@ class TextInputWidget extends InputWidget {
 		if ( isset( $config['multiline'] ) && $config['multiline'] ) {
 			return 'multiline';
 		} else {
-			$type = in_array( $config['type'], array( 'text', 'password', 'search', 'email', 'url' ) ) ?
+			$type = in_array( $config['type'], [ 'text', 'password', 'search', 'email', 'url' ] ) ?
 				$config['type'] :
 				'text';
 			return $type;
