@@ -34,10 +34,10 @@ class TagTest extends PHPUnit_Framework_TestCase {
 	 */
 	public function testClasses() {
 		$tag = new Tag();
-		$tag->addClasses( array( 'class1', 'class2' ) );
+		$tag->addClasses( [ 'class1', 'class2' ] );
 		$this->assertTrue( $tag->hasClass( 'class1' ) );
 		$this->assertTrue( $tag->hasClass( 'class2' ) );
-		$tag->removeClasses( array( 'class1' ) );
+		$tag->removeClasses( [ 'class1' ] );
 		$this->assertFalse( $tag->hasClass( 'class1' ) );
 		$this->assertTrue( $tag->hasClass( 'class2' ) );
 	}
@@ -49,14 +49,14 @@ class TagTest extends PHPUnit_Framework_TestCase {
 	 */
 	public function testAttributes() {
 		$tag = new Tag();
-		$tag->setAttributes( array(
+		$tag->setAttributes( [
 			'attr1' => 'foo',
 			'attr2' => 'bar',
-		) );
+		] );
 		$this->assertEquals( 'foo', $tag->getAttribute( 'attr1' ) );
 		$this->assertEquals( 'bar', $tag->getAttribute( 'attr2' ) );
 		$this->assertEquals( null, $tag->getAttribute( 'attr3' ) );
-		$tag->removeAttributes( array( 'attr1' ) );
+		$tag->removeAttributes( [ 'attr1' ] );
 		$this->assertEquals( null, $tag->getAttribute( 'attr1' ) );
 		$this->assertEquals( 'bar', $tag->getAttribute( 'attr2' ) );
 	}
@@ -68,7 +68,7 @@ class TagTest extends PHPUnit_Framework_TestCase {
 		$tag = new Tag();
 		$tag2 = new Tag();
 		$tag3 = new Tag();
-		$tag->setAttributes( array( 'id' => 'foobar' ) );
+		$tag->setAttributes( [ 'id' => 'foobar' ] );
 		$tag->ensureInfusableId();
 		$tag2->ensureInfusableId();
 		$tag3->ensureInfusableId();
@@ -90,20 +90,20 @@ class TagTest extends PHPUnit_Framework_TestCase {
 	 * @note: Keep tests in sync with core.test.js
 	 */
 	public static function provideIsSafeUrl() {
-		return array(
-			array( 'javascript:evil();', false ),
-			array( 'foo:bar', false ),
-			array( 'relative.html', false ),
-			array( '', true ),
-			array( 'http://example.com/', true ),
-			array( '//example.com/', true ),
-			array( '/', true ),
-			array( '..', false ),
-			array( '?foo=bar', true ),
-			array( '#top', true ),
-			array( '/relative', true ),
-			array( './relative', true ),
-			array( '/wiki/Extra:Colon', true ),
-		);
+		return [
+			[ 'javascript:evil();', false ],
+			[ 'foo:bar', false ],
+			[ 'relative.html', false ],
+			[ '', true ],
+			[ 'http://example.com/', true ],
+			[ '//example.com/', true ],
+			[ '/', true ],
+			[ '..', false ],
+			[ '?foo=bar', true ],
+			[ '#top', true ],
+			[ '/relative', true ],
+			[ './relative', true ],
+			[ '/wiki/Extra:Colon', true ],
+		];
 	}
 }
